rootston: move part of desktop.c to view.c, use an interface for views
