+++ rootston/meson.build
@@ -12,10 +12,11 @@ sources = [
 	'seat.c',
 	'switch.c',
 	'text_input.c',
+	'view.c',
 	'virtual_keyboard.c',
 	'wl_shell.c',
-	'xdg_shell.c',
 	'xdg_shell_v6.c',
+	'xdg_shell.c',
 ]
 
 if conf_data.get('WLR_HAS_XWAYLAND', 0) == 1
